Extracted Venue in own module
