+++ app/src/test/java/net/squanchy/service/firebase/model/schedule/FirestoreTracksRepositoryTest.kt
@@ -3,8 +3,8 @@ package net.squanchy.service.firebase.model.schedule
 import io.reactivex.Observable
 import net.squanchy.schedule.domain.view.aTrack
 import net.squanchy.schedule.firestore.FirestoreTracksRepository
-import net.squanchy.service.firebase.FirestoreDbService
 import net.squanchy.schedule.firestore.aFirestoreTrack
+import net.squanchy.service.firebase.FirestoreDbService
 import net.squanchy.support.checksum.Checksum
 import org.junit.Before
 import org.junit.Rule
