+++ firebase/firestore-db-service/src/test/java/net/squanchy/service/firebase/model/conferenceinfo/FirestoreVenueMapperTest.kt
@@ -1,9 +1,6 @@
-package net.squanchy.service.firebase
+package net.squanchy.service.firebase.model.conferenceinfo
 
-import com.google.common.truth.Truth.assertThat
 import com.google.firebase.firestore.GeoPoint
-import net.squanchy.schedule.firestore.aFirestoreVenue
-import net.squanchy.venue.domain.view.toVenue
 import org.junit.Test
 
 private const val A_STRING = "Whatever"
