+++ schedule/schedule-service-firestore/src/test/java/net/squanchy/schedule/firestore/FirestoreEventMapperTest.kt
@@ -1,16 +1,10 @@
-package net.squanchy.service.firebase
+package net.squanchy.schedule.firestore
 
 import com.google.common.truth.Truth.assertThat
 import net.squanchy.A_DATE
 import net.squanchy.A_TIMEZONE
 import net.squanchy.schedule.domain.view.Event
 import net.squanchy.schedule.domain.view.ExperienceLevel
-import net.squanchy.schedule.firestore.aFirestoreEvent
-import net.squanchy.schedule.firestore.aFirestorePlace
-import net.squanchy.schedule.firestore.aFirestoreTrack
-import net.squanchy.schedule.firestore.toEvent
-import net.squanchy.schedule.firestore.toPlace
-import net.squanchy.schedule.firestore.toTrack
 import net.squanchy.support.checksum.Checksum
 import net.squanchy.support.lang.getOrThrow
 import org.junit.Before
