+++ schedule/schedule-service-firestore/src/test/java/net/squanchy/schedule/firestore/FirestoreFixtures.kt
@@ -73,22 +73,6 @@ fun aFirestoreDay(
     this.date = date
 }
 
-fun aFirestoreVenue(
-    name: String = "Venue Of The Conference",
-    address: String = "Conference rd",
-    latLon: GeoPoint = GeoPoint(0.0, 0.0),
-    description: String = "This is the conference you're looking for",
-    mapUrl: String = "www.maps.google.com/conference",
-    timezone: String = "Europe/Rome"
-) = FirestoreVenue().apply {
-    this.name = name
-    this.address = address
-    this.latLon = latLon
-    this.description = description
-    this.mapUrl = mapUrl
-    this.timezone = timezone
-}
-
 fun aFirestoreEvent(
     id: String = "banana",
     title: String = "Hello \uD83C\uDF4C", // Yes, that's a banana emoji. You never know
