+++ schedule/schedule-service-firestore/src/test/java/net/squanchy/schedule/firestore/FirestoreSpeakerMapperTest.kt
@@ -1,8 +1,6 @@
-package net.squanchy.service.firebase
+package net.squanchy.schedule.firestore
 
 import com.google.common.truth.Truth.assertThat
-import net.squanchy.schedule.firestore.aFirestoreSpeaker
-import net.squanchy.schedule.firestore.toSpeaker
 import net.squanchy.support.checksum.Checksum
 import net.squanchy.support.lang.getOrThrow
 import org.junit.Before
