+++ venue/venue-domain/src/main/java/net/squanchy/venue/domain/view/Venue.kt
@@ -1,6 +1,5 @@
 package net.squanchy.venue.domain.view
 
-import net.squanchy.service.firebase.model.conferenceinfo.FirestoreVenue
 import org.joda.time.DateTimeZone
 
 @Suppress("LongParameterList") // This is just a big model - TODO refactor this to split it up
@@ -12,14 +11,4 @@ data class Venue(
     val description: String,
     val mapUrl: String,
     val timeZone: DateTimeZone
-)
-
-fun FirestoreVenue.toVenue() = Venue(
-    name = name,
-    address = address,
-    latitude = latLon.latitude,
-    longitude = latLon.longitude,
-    description = description,
-    mapUrl = mapUrl,
-    timeZone = DateTimeZone.forID(timezone)
 )
